Rearrangement of KroneckerDelta class
Merged duplicated KroneckerDelta classes in sympy/physics/quantum/kronockerdelta and sympy/physics/secondquant. The merged KroneckerDelta class is placed in sympy/functions/special/tensor_functions.py and replaces the original Dij classes in that module. The corresponding tests and imports have also been adjusted.

GCI Task Link: http://www.google-melange.com/gci/task/view/google/gci2011/7121285
